+++ app/bridge_app/src/entity/token_relations.rs
@@ -6,9 +6,9 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "token_relations")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub evm_token_id: i32,
-    pub solana_token_id: i32,
+    pub id: i64,
+    pub evm_token_id: i64,
+    pub solana_token_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
